"fixed" bug #1 - DHCP Renew
Added a 5 second sleep time and force a second renew on the interface
who's ip has changed. Not entirely sure why this needs to happen, script
gets a little bit funky for a few seconds but does recover to a normal
state after a few seconds. Still need to find out why this happens but
this band-aid should work for now...